implementations/feature5_support_chat/requirements.txt
Flask==2.2.5
flask-socketio==5.3.2
python-socketio==5.4.0
simple-websocket==0.5.2
psycopg2-binary
python-dotenv==0.20.0